Add submission and voting of problems
This is a rather large change. It adds a number of features including
the submission of problems along with a markdown editor and previewing
abilites, as well as a auto-completing tag editor. Also voting for
problems and view counting is implemented, however some sort of control
is needed to prevent abuse.