Switch to a simpler post model
Problems and solutions are simplified to only contain one body attribute
that contains the code that is to be/has been refactored. This not only
simplifies the submission process but also greatly streamlines the
application.

Instead of MarkItUp, Ace is used as the editor, and all of the previous
prettify code highlighting has also been replaces by ace.